+++ CHANGELOG.md
@@ -4,6 +4,23 @@ This project is following [Semantic Versioning](http://semver.org)
 
 ## [Unreleased][]
 
+## [0.3.0][] - 2018-11-23
+
+### Changed
+
+ - use @deskpro/apps-sdk@0.11.0
+ - use @deskpro/apps-components@0.11.0
+ - use @deskpro/apps-installer@1.0.8
+ - use @deskpro/apps-react-scripts@0.5.0
+
+## [0.2.6][] - 2018-08-16
+
+### Changed
+
+ - using the new apps structure and sdks
+ - use @deskpro/apps-sdk@0.5.0
+ - use @deskpro/apps-components@0.5.0
+
 ## [0.2.5][] - 2018-04-30
 
 ### Added
@@ -28,7 +45,9 @@ This project is following [Semantic Versioning](http://semver.org)
     - travis will attach builds to Github PR's when enabled via s3 environment variables
 
 
-[Unreleased]: https://github.com/DeskproApps/custom-html/compare/v0.2.5...HEAD
+[Unreleased]: https://github.com/DeskproApps/custom-html/compare/v0.3.0...HEAD
+[0.3.0]: https://github.com/DeskproApps/custom-html/compare/v0.2.6...v0.3.0
+[0.2.6]: https://github.com/DeskproApps/custom-html/compare/v0.2.5...v0.2.6
 [0.2.5]: https://github.com/DeskproApps/custom-html/compare/v0.2.4...v0.2.5
 [0.2.4]: https://github.com/DeskproApps/custom-html/compare/v0.2.3...v0.2.4
 [0.2.3]: https://github.com/DeskproApps/custom-html/tree/v0.2.3